Move oonirun router to v2
