+++ ooniapi/services/oonirun/tests/test_oonirun.py
@@ -7,7 +7,7 @@ from datetime import datetime, timedelta, timezone
 import time
 
 from oonirun import models
-from oonirun.routers.oonirun import utcnow_seconds
+from oonirun.routers.v2 import utcnow_seconds
 import pytest
 
 import sqlalchemy as sa
